% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_aggregation.R
\name{interval_aggregation}
\alias{interval_aggregation}
\title{Aggregate the raw data into an equally-spaced data frame}
\usage{
interval_aggregation(data, variable, ID, timestamp,
  format = "\%Y-\%m-\%dT\%H:\%M:\%S", interval = 60,
  aggregatefun = sum, is.na.zero = TRUE, remove.na = TRUE)
}
\arguments{
\item{data}{Data in a data.frame format}

\item{variable}{Name of variable in the data set you would like to re-structure}

\item{ID}{Name of ID variable in the data set}

\item{timestamp}{Name of time variable in the data set}

\item{format}{Date/time format of timestamp. Default is "%Y-%m-%dT%H:%M:%S" (i.e. yyyy-mm-dd hh:mm:ss).}

\item{interval}{The interval of time in minute. Default is 60 (i.e. The measurement is evaluated every hour.)}

\item{aggregatefun}{Function used for aggregation. Default is sum.}

\item{is.na.zero}{If TRUE, NA is considered as 0. If FALSE, the NA is considered as missing. Defulat is TRUE.}

\item{remove.na}{If TRUE, the NAs will be removed in the aggregated data. Defulat is TRUE.}
}
\description{
Aggregate the raw data into an equally-spaced data frame
}
