% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_plot.R
\name{raw_plot}
\alias{raw_plot}
\title{Plot the aggregated data per individual}
\usage{
raw_plot(data, variable, ID, Date, Time)
}
\arguments{
\item{data}{Data in a data.frame format (e.g. Output of function 'interval_aggregation')}

\item{variable}{Name of variable in the data set you would like to plot}

\item{ID}{Name of ID variable in the data set}

\item{Date}{Name of date variable in the data set}

\item{Time}{Name of time variable in the data set}
}
\description{
Plot the aggregated data per individual
}
